backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hardware-scanner-api"
version = "1.0.0"
description = "Hardware Inventory Scanner FastAPI Backend"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "groq",
    "python-multipart",
    "openpyxl",
    "python-dotenv",
]
